offers from firebase

+++ lib/models/offer_model.dart
@@ -3,7 +3,7 @@ import 'package:pharmacyapp/models/drug_model.dart';
 class OfferItem {
   Drug drug;
   bool percentage = true;
-  int offer;
+  double offer;
 
   OfferItem(this.drug, this.offer, {bool isPercentage = true}) {
     percentage = isPercentage;
